application.js の記述をcalendar.js ファイルに切り出し
